🏢 Refactor editables into their own file
